+++ packages/backend/src/controllers/sqlRunnerController.ts
@@ -2,6 +2,7 @@ import {
     ApiErrorPayload,
     ApiJobScheduledResponse,
     ApiWarehouseCatalog,
+    ApiWarehouseTableFields,
     SqlRunnerBody,
 } from '@lightdash/common';
 import {
@@ -63,7 +64,7 @@ export class SqlRunnerController extends BaseController {
         @Path() projectUuid: string,
         @Path() tableName: string,
         @Request() req: express.Request,
-    ): Promise<ApiWarehouseCatalog> {
+    ): Promise<ApiWarehouseTableFields> {
         this.setStatus(200);
 
         return {