+++ packages/frontend/src/features/notifications/hooks/useNotifications.tsx
@@ -24,7 +24,7 @@ export const useGetNotifications = (
         ['notifications', type],
         () => getNotifications(type),
         {
-            refetchInterval: 30 * 1000, // 30 seconds
+            refetchInterval: 3 * 60 * 1000, // 3 minutes
             retry: (_, error) => error.error.statusCode !== 403,
             enabled,
         },